add_CellLine_annotation

add cellline annotation to a data.table with metadata

Usage

add_CellLine_annotation(
  dt_metadata,
  DB_cellid_header = "cell_line_identifier",
  DB_cell_annotate = c("cell_line_name", "primary_tissue", "doubling_time",
    "parental_identifier", "subtype"),
  fname = "cell_lines.csv",
  fill = "unknown",
  annotationPackage = if ("gDRinternal" %in% .packages(all.available = TRUE)) {
    
    "gDRinternal"
 } else {
     "gDRtestData"
 },
  externalSource = Sys.getenv("GDR_CELLLINE_ANNOTATION")
)

Arguments

dt_metadata

data.table with metadata

DB_cellid_header

string with colnames with cell line identifier in the annotation file

DB_cell_annotate

character vector with mandatory colnames used in the annotation file

fname

string with file name with annotation

fill

string indicating how unknown cell lines should be filled in the DB

annotationPackage

string indication name of the package containing cellline annotation

externalSource

string with path to external file with annotation data; by default it checks 'GDR_CELLLINE_ANNOTATION' env var. This file should contain columns such as gnumber, drug_name and drug_moa

Value

data.table with metadata with annotated cell lines

Details

The logic of adding celline annotation for dt_metadata based on the annotation file stored in gDRtestData. Other fields are set as "unknown". This approach will be corrected once we will implement final solution for adding cell lines.

Examples

add_CellLine_annotation(
  data.table::data.table(
    clid = "123", 
    CellLineName = "name of the cell line")
)
#> INFO [2024-04-23 20:40:04] Merge with Cell line info
#>      clid CellLineName  Tissue ReferenceDivisionTime parental_identifier
#>    <char>       <char>  <char>                 <num>              <char>
#> 1:    123          123 unknown                    NA                 123
#>    subtype        i.CellLineName
#>     <char>                <char>
#> 1: unknown name of the cell line